Enhanced resources support.
Added support for copying all types of files referenced in the articles.
Better handling of resources files (with the idea that the template
directory contains only the html and css basic files, all other (fonts,
javascript, favicons) being stored in the resources. This should allow
quick modification/upload of CSS files for instance.